requirements.txt
# Core Dependencies - Production Pinned Versions
langchain==0.2.12
langchain-groq==0.1.6
langchain-community==0.2.9
langchain-core==0.2.27
# Vector Database & Embeddings
qdrant-client==1.15.1  # Latest release with Python 3.12+ support
sentence-transformers==3.0.1
PyMuPDF==1.24.7  # Stable, with Python 3.13 wheels on macOS
python-docx==0.8.11
tiktoken==0.7.0
faiss-cpu==1.8.0.post1
# Enhanced LLM & Structured Output
pydantic==2.8.2
pydantic-settings==2.3.3
jsonschema==4.23.0
# Advanced Text Processing
nltk==3.9.1
textstat==0.7.4
readability==0.3.2
networkx==3.3
spacy==3.7.2
# Data Processing & Analysis
pandas==2.2.2
openpyxl==3.1.5
numpy==1.26.4
scikit-learn==1.5.1
# UI & Visualization
streamlit==1.51.0
plotly==5.24.1
# RAGFlow Enhanced Features
unstructured==0.18.14
unstructured[pdf]==0.18.14 ; platform_system != "Windows"
layoutparser==0.3.4
# Production Monitoring
prometheus-client==0.21.0
structlog==24.4.0
# Testing
pytest==8.3.2
pytest-cov==5.0.0
pytest-asyncio==0.24.0
# Utils
python-dotenv==1.0.1
retry==0.9.2
asyncio-throttle==1.0.2
tqdm==4.66.4
# HTTP & API
httpx==0.24.1
fastapi==0.101.1
uvicorn==0.24.0
